Clean up extensions
