+++ Client/Extensions/CGRect+Utilies.swift
@@ -1,0 +1,22 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0
+
+import UIKit
+
+extension CGRect {
+    var center: CGPoint {
+        get {
+            return CGPoint(x: size.width / 2, y: size.height / 2)
+        }
+        set {
+            self.origin = CGPoint(x: newValue.x - size.width / 2, y: newValue.y - size.height / 2)
+        }
+    }
+
+    func visibilityRatio(relativeTo otherFrame: CGRect) -> CGFloat {
+        let inter = self.intersection(otherFrame)
+        let ratio = (inter.width * inter.height) / (self.width * self.height)
+        return ratio
+    }
+}
