+++ Client/Extensions/NSAttributedString+Utilies.swift
