--- Client/Extensions/SnapKitExtensions.swift
@@ -1,13 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0
-
-import Foundation
-import SnapKit
-
-extension UIView {
-
-    var safeArea: ConstraintBasicAttributesDSL {
-        return self.safeAreaLayoutGuide.snp
-    }
-}
