+++ Client/Extensions/UIImageView+Utilies.swift
@@ -7,15 +7,6 @@ import Storage
 import Shared
 import Kingfisher
 
-extension UIColor {
-    func image(_ size: CGSize = CGSize(width: 1, height: 1)) -> UIImage {
-        return UIGraphicsImageRenderer(size: size).image { rendererContext in
-            self.setFill()
-            rendererContext.fill(CGRect(origin: .zero, size: size))
-        }
-    }
-}
-
 public extension UIImageView {
 
     func setImageAndBackground(forIcon icon: Favicon?, website: URL?, completion: @escaping () -> Void) {
@@ -81,31 +72,5 @@ public extension UIImageView {
         let templateImage = self.image?.withRenderingMode(.alwaysTemplate)
         self.image = templateImage
         self.tintColor = color
-    }
-}
-
-extension UIImage {
-    func overlayWith(image: UILabel) -> UIImage {
-        UIGraphicsBeginImageContextWithOptions(CGSize(width: size.width, height: size.height), false, 0.0)
-        draw(in: CGRect(origin: CGPoint.zero, size: size))
-        image.draw(CGRect(origin: CGPoint.zero, size: image.frame.size))
-        let newImage = UIGraphicsGetImageFromCurrentImageContext()!
-        UIGraphicsEndImageContext()
-
-        return newImage
-    }
-
-    func overlayWith(image: UIImage) -> UIImage {
-        let newSize = CGSize(width: size.width, height: size.height)
-        UIGraphicsBeginImageContextWithOptions(newSize, false, 0.0)
-        draw(in: CGRect(origin: CGPoint.zero, size: newSize))
-        let modifier: CGFloat = 0.35
-        image.draw(in: CGRect(origin: CGPoint(x: 15, y: 16),
-                              size: CGSize(width: size.width * modifier,
-                                           height: size.height * modifier)))
-        let newImage = UIGraphicsGetImageFromCurrentImageContext()!
-        UIGraphicsEndImageContext()
-
-        return newImage
     }
 }
