+++ Client/Extensions/UIPasteboard+Utilies.swift
