+++ Client/Extensions/URL+Mail.swift
