+++ Client/UIWindow+Utilies.swift
